fix: Disable opening notifications on Android U+
This commit disables the "Open notification" functionality for Android versions higher than Tiramisu (API 33). This is a temporary measure due to changes in how PendingIntents are handled in newer Android versions, which currently prevents the feature from working as expected.

The visibility of the "Open notification" text view (`tvOpenNotification`) in the `AppNotificationAdapter` is now conditional on the Android version.